% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modeling.R
\name{change_output_col_pos}
\alias{change_output_col_pos}
\title{Function places the dependent variable as the first/last column in the data will make things easier}
\usage{
change_output_col_pos(data, output_col, position = "first")
}
\arguments{
\item{output_col}{column to reorder}

\item{position}{first or last}
}
\value{
data with output_col as position defines
}
\description{
Remember to keep only the variables that can be used for building model
}
