% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{convert_all_numeric}
\alias{convert_all_numeric}
\title{Function that converts all of the variables into numeric type}
\usage{
convert_all_numeric(data)
}
\description{
Function that converts all of the variables into numeric type
}
