% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{count_per_group}
\alias{count_per_group}
\title{Function that counts number of rows and percentage based on group/s}
\usage{
count_per_group(data, ...)
}
\arguments{
\item{data}{name of the dataframe/datatable}

\item{...}{a column name or a vector with multiple names}
}
\description{
Function that counts number of rows and percentage based on group/s
}
