% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{encoding_col}
\alias{encoding_col}
\title{Function that converts the given columns' values into individual column with TRUE or FALSE}
\usage{
encoding_col(data, column)
}
\arguments{
\item{...}{list of variables to encode as dummy variable}
}
\value{
one hot encode variables as logical values to avoid possible confusion as number
}
\description{
Function that converts the given columns' values into individual column with TRUE or FALSE
}
