% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{factor_func}
\alias{factor_func}
\title{Function that converts all of the variables listed into factor type}
\usage{
factor_func(data, ...)
}
\arguments{
\item{...}{list of variables to factorize}
}
\description{
Function that converts all of the variables listed into factor type
}
