% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modeling.R
\name{glm_model}
\alias{glm_model}
\title{Function to print logistic regression model using formula}
\usage{
glm_model(formula)
}
\arguments{
\item{formula}{output ~ input}
}
\value{
Logistic regression model
}
\description{
Function to print logistic regression model using formula
}
