% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file.R
\name{install_packages}
\alias{install_packages}
\title{Function to load packages already installed otherwise install then load, may require rJava if not already installed}
\usage{
install_packages(pkg)
}
\arguments{
\item{pkg}{name of packages as a vector}
}
\value{
check if packages are already installed, install necessary packages and load them
}
\description{
Function to load packages already installed otherwise install then load, may require rJava if not already installed
}
