% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{is_date}
\alias{is_date}
\title{compute basic statistics for date types}
\usage{
is_date(x)
}
\description{
first of all find all of the columns that are dates
r doesn't have a simple function for that so we will build our own function
}
