% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{join_summary}
\alias{join_summary}
\title{Function combines the summary of date and numeric variables}
\usage{
join_summary(data)
}
\arguments{
\item{data}{name}
}
\description{
Function combines the summary of date and numeric variables
}
