% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file.R
\name{load_datatable}
\alias{load_datatable}
\title{Function to load a file: replaces blank space by NA using na.string and check if there are duplicate column names}
\usage{
load_datatable(file, check.names = TRUE)
}
\arguments{
\item{file}{name of file along with its location}

\item{check.names}{should names starting with a number, or that are duplicated be treated? TRUE or FALSE}
}
\value{
loads the data as data.table
}
\description{
Function to load a file: replaces blank space by NA using na.string and check if there are duplicate column names
}
