% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{load_excel}
\alias{load_excel}
\title{Function to load excel file: replaces blank space by NA using na.string and check if there are duplicate column names}
\usage{
load_excel(file, check.names = FALSE)
}
\description{
Function to load excel file: replaces blank space by NA using na.string and check if there are duplicate column names
}
