% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{log_odds_prob}
\alias{log_odds_prob}
\title{log odds to odds
confidence interval of predictors
probability and 95% CI using standard errors}
\usage{
log_odds_prob(model)
}
\description{
log odds to odds
confidence interval of predictors
probability and 95% CI using standard errors
}
