% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modeling.R
\name{make_formula}
\alias{make_formula}
\title{This function loads packages without warnings}
\usage{
make_formula(data, index)
}
\arguments{
\item{pkg}{name of packages as a vector}

\item{join_summary}{uses summary_numbers and summary_dates to give a tidy summary}
}
\value{
distinct numbers, and basic statistics for each variable in the data
}
\description{
This function loads packages without warnings
}
