% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modeling.R
\name{make_formula_1var}
\alias{make_formula_1var}
\title{Function converts text format into formula format}
\usage{
make_formula_1var(data)
}
\arguments{
\item{data}{after passing trans_vars_func()}
}
\value{
list of formulae with one-one pair of each independent variables with the dependent variable
}
\description{
Function converts text format into formula format
}
