% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{plot_rcor}
\alias{plot_rcor}
\title{Correlation can be evaluated only for numeric type
Plot the correlationship
Insignificant correlations are leaved blank}
\usage{
plot_rcor(data)
}
\description{
Correlation can be evaluated only for numeric type
Plot the correlationship
Insignificant correlations are leaved blank
}
