% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file.R
\name{quietly_packages}
\alias{quietly_packages}
\title{Function to load packages without warnings}
\usage{
quietly_packages(pkg, suppress = "both")
}
\arguments{
\item{pkg}{name of packages as a vector}

\item{suppress}{warnings, messages or both}
}
\description{
Function to load packages without warnings
}
