% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{remove_duplicate_rows}
\alias{remove_duplicate_rows}
\title{Function that removes any duplicate rows}
\usage{
remove_duplicate_rows(data)
}
\arguments{
\item{data}{name}
}
\value{
data without duplicated rows
}
\description{
Function that removes any duplicate rows
}
