% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{replace_NA}
\alias{replace_NA}
\title{Function to replace "NA" because NA as string is not treated as NA}
\usage{
replace_NA(data, set_char = "missing")
}
\arguments{
\item{set_char}{replace NULL values in "NA" by set_char}
}
\description{
Function to replace "NA" because NA as string is not treated as NA
}
