% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modeling.R
\name{single_combn}
\alias{single_combn}
\title{This function produces all combinations of variables with the dependent variable}
\usage{
single_combn(data)
}
\value{
keep the dependent variable as the first column
}
\description{
This function produces all combinations of variables with the dependent variable
}
