% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{summary_dates}
\alias{summary_dates}
\title{Function compute basic statistics for date types}
\usage{
summary_dates(data)
}
\arguments{
\item{data}{name}
}
\value{
summary of date variables
}
\description{
find columns that are dates using is_date() then covert all of the dates into one format for consistency. We can't use ":" to separate like in summary_numbers() because of %H:%M:%S in the column. Therefore we are using the regular expression to split using the first occurance of ":" and keeping the remaining values
}
