% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{summary_dcount}
\alias{summary_dcount}
\title{Function count the number of unique values in each variable}
\usage{
summary_dcount(data)
}
\arguments{
\item{data}{name}
}
\value{
count of distinct values for each variable in the data
}
\description{
Function count the number of unique values in each variable
}
