% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{summary_logicals}
\alias{summary_logicals}
\title{Function counts the number of TRUE and FALSE for logical type}
\usage{
summary_logicals(data)
}
\arguments{
\item{data}{name}
}
\description{
Function counts the number of TRUE and FALSE for logical type
}
