% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{summary_numbers}
\alias{summary_numbers}
\title{Function compute basic statistics for numeric types}
\usage{
summary_numbers(data)
}
\arguments{
\item{data}{name}
}
\value{
TRUE or FALSE whether variable is of date type
}
\description{
Function compute basic statistics for numeric types
}
