% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{tidy_packages}
\alias{tidy_packages}
\title{Load necessary packages
This function loads packages without warnings}
\usage{
tidy_packages(pkg)
}
\arguments{
\item{pkg}{name of packages as a vector}

\item{join_summary}{uses summary_numbers and summary_dates to give a tidy summary}
}
\value{
distinct numbers, and basic statistics for each variable in the data
}
\description{
Load necessary packages
This function loads packages without warnings
}
\examples{
tidy_packages(packages)
}
