% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{tidy_summary}
\alias{tidy_summary}
\title{Function to print tidy version of "summary" function}
\usage{
tidy_summary(data)
}
\arguments{
\item{data}{name}
}
\value{
distinct numbers, and basic statistics for each variable in the data. characters and logical data types don't have min, median, etc. which can be replaced with a dash "-" instead of NAs
}
\description{
Function to print tidy version of "summary" function
}
