% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modeling.R
\name{trans_vars_func}
\alias{trans_vars_func}
\title{This function creates a column with formula for the dependent variable with each of the independent variable}
\usage{
trans_vars_func(data)
}
\value{
transpose the table created by single_combn then add a new column that has output ~ input structure
}
\description{
This function creates a column with formula for the dependent variable with each of the independent variable
}
