% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{unique_keys_more}
\alias{unique_keys_more}
\title{Function that prints distinct values based on more than one group}
\usage{
unique_keys_more(data, ...)
}
\arguments{
\item{data}{name of the dataframe/datatable}

\item{...}{a column name or a vector with multiple names}
}
\description{
Function that prints distinct values based on more than one group
}
