% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{unique_keys_one}
\alias{unique_keys_one}
\title{Function that prints distinct values based on one variable}
\usage{
unique_keys_one(data, ...)
}
\arguments{
\item{data}{name of the dataframe/datatable}

\item{...}{a column name or a vector with multiple names}
}
\value{
print as a vector instead of a column format
}
\description{
Function that prints distinct values based on one variable
}
