% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_cleaning.R
\name{z_score_transform}
\alias{z_score_transform}
\title{Function standarizes all the independent variables (convert the mean into 0 with standard deviation 1) using scale() function, which by default does z-score transformation}
\usage{
z_score_transform(data, ...)
}
\arguments{
\item{...}{list of variables to factorize}
}
\description{
Function standarizes all the independent variables (convert the mean into 0 with standard deviation 1) using scale() function, which by default does z-score transformation
}
