% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{handle_NA}
\alias{handle_NA}
\title{We must impute the missing data in a variable
We can remove the entire row (avoid if relatively small data size), replace with either 0 or mean/median value of all the available variable
It is okay to replace NA with 0 when counting the values
replace NAs with 0 for integers and replace NAs with "missing" in character columns}
\usage{
handle_NA(data)
}
\description{
We must impute the missing data in a variable
We can remove the entire row (avoid if relatively small data size), replace with either 0 or mean/median value of all the available variable
It is okay to replace NA with 0 when counting the values
replace NAs with 0 for integers and replace NAs with "missing" in character columns
}
