% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{install_packages}
\alias{install_packages}
\title{Function to load packages already installed otherwise install then load}
\usage{
install_packages(pkg)
}
\arguments{
\item{pkg}{name of packages as a vector}
}
\value{
install necessary packages and load them
}
\description{
Function to load packages already installed otherwise install then load
}
\examples{
install_packages(packages)
check if packages are already installed
load the packges
}
