% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{is_date}
\alias{is_date}
\title{Function checks whether a variable in the data is of date type or not. r doesn't have a simple function for that so we will build our own function}
\usage{
is_date(x)
}
\arguments{
\item{x}{variable as one-dimensional array}
}
\value{
TRUE or FALSE whether variable is of date type
}
\description{
Function checks whether a variable in the data is of date type or not. r doesn't have a simple function for that so we will build our own function
}
