% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{join_summary}
\alias{join_summary}
\title{combine both dates and numbers
string with . at end gives odd results when removing whitespace
Order some of the columns}
\usage{
join_summary(data)
}
\description{
combine both dates and numbers
string with . at end gives odd results when removing whitespace
Order some of the columns
}
