% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{load_datatable}
\alias{load_datatable}
\title{install.packages(c("devtools", "roxygen2"))
tidy_packages(c("devtools", "roxygen2"))
roxygen2 package provides an in-source documentation system that automatically generates R documentation (Rd) files
setwd(parentDirectory)
devtools::create("easyDataScience")
redo document() to update as needed
setwd("./easyDataScience")
Run devtools::document() to convert roxygen comments to .Rd files.
setwd("..")
install("easyDataScience")
easyDataScience environment that is the parent environment to the global environment.
search()
Load data}
\usage{
load_datatable(file)
}
\arguments{
\item{file}{name of file along with its location}
}
\value{
loads the data
}
\description{
install.packages(c("devtools", "roxygen2"))
tidy_packages(c("devtools", "roxygen2"))
roxygen2 package provides an in-source documentation system that automatically generates R documentation (Rd) files
setwd(parentDirectory)
devtools::create("easyDataScience")
redo document() to update as needed
setwd("./easyDataScience")
Run devtools::document() to convert roxygen comments to .Rd files.
setwd("..")
install("easyDataScience")
easyDataScience environment that is the parent environment to the global environment.
search()
Load data
}
\examples{
df <- load_datatable("~/Documents/data.csv")
Function to load a file: replaces blank space by NA using na.string and check if there are duplicate column names
}
