% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_file.R
\name{load_excel}
\alias{load_excel}
\title{Function to load excel file: check if there are duplicate column names and find columns that are dates}
\usage{
load_excel(file, check.names = FALSE, detectDates = TRUE)
}
\arguments{
\item{file}{name of excel file along with its location}
}
\value{
loads the data as excel
}
\description{
Function to load excel file: check if there are duplicate column names and find columns that are dates
}
