% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modeling.R
\name{log_odds_prob}
\alias{log_odds_prob}
\title{Function takes logistic model and converts the coeficients in probability scale}
\usage{
log_odds_prob(model)
}
\arguments{
\item{model}{logistic regression model}
}
\value{
probability of coefficients of independent variables in logistic model
}
\description{
Function takes logistic model and converts the coeficients in probability scale
}
