% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modeling.R
\name{make_formula_1var}
\alias{make_formula_1var}
\title{This function write a logistic regression model}
\usage{
make_formula_1var(data)
}
\arguments{
\item{pkg}{name of packages as a vector}

\item{join_summary}{uses summary_numbers and summary_dates to give a tidy summary}
}
\value{
distinct numbers, and basic statistics for each variable in the data
}
\description{
This function write a logistic regression model
}
