% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{odds_coef}
\alias{odds_coef}
\title{Logistic regression coefficients give the change in the log odds of the outcome for a one unit increase in the predictor variable.
log odds to odds
confidence interval of predictors
odds ratios and 95% CI using standard errors}
\usage{
odds_coef(model)
}
\description{
Logistic regression coefficients give the change in the log odds of the outcome for a one unit increase in the predictor variable.
log odds to odds
confidence interval of predictors
odds ratios and 95% CI using standard errors
}
