% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modeling.R
\name{odds_coef}
\alias{odds_coef}
\title{Function takes logistic model and converts the coeficients in odds scale. Logistic regression model's coefficients give the change in the log odds of the outcome for a one unit increase in the predictor variable.}
\usage{
odds_coef(model)
}
\arguments{
\item{model}{logistic regression model}
}
\value{
odds of coefficients of independent variables in logistic model
}
\description{
Function takes logistic model and converts the coeficients in odds scale. Logistic regression model's coefficients give the change in the log odds of the outcome for a one unit increase in the predictor variable.
}
