% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modeling.R
\name{plot_rcor}
\alias{plot_rcor}
\title{This function plots the correlationship between variables, insignificant correlations are leaved blank}
\usage{
plot_rcor(data)
}
\value{
Correlationship between numeric variables
}
\description{
This function plots the correlationship between variables, insignificant correlations are leaved blank
}
