% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{replace_NA}
\alias{replace_NA}
\title{we need another function to replace "NA" because NA as string is not treated as NA}
\usage{
replace_NA(data)
}
\description{
we need another function to replace "NA" because NA as string is not treated as NA
}
