% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{save_data}
\alias{save_data}
\title{This function saves the data}
\usage{
save_data(data, file_location)
}
\arguments{
\item{data}{write the name of data}

\item{file_location}{write the location where to save the data and in which format}
}
\value{
saves data as the format defined in file_location
}
\description{
This function saves the data
}
\examples{
save the data after data wrangling and manipulation: '~/Documents/[folders]/[file_name.csv]'

}
