% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{single_combn}
\alias{single_combn}
\title{All combinations of variables with the dependent variable
keep the dependent variable as the first column}
\usage{
single_combn(data)
}
\description{
All combinations of variables with the dependent variable
keep the dependent variable as the first column
}
