% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_modeling.R
\name{single_combn}
\alias{single_combn}
\title{This function produces all combinations of variables with the dependent variable}
\usage{
single_combn(data)
}
\value{
create dataframe with nrow 2 and ncol ncol(df) - 1 such that the first row in each column is the dependent variable and the second row in each column is one of the independent variables
}
\description{
Remember to keep the dependent variable as the first column
}
