% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration.R
\name{summary_dates}
\alias{summary_dates}
\title{Function compute basic statistics for date types}
\usage{
summary_dates(data)
}
\arguments{
\item{data}{name}
}
\value{
TRUE or FALSE whether variable is of date type
}
\description{
Function compute basic statistics for date types
}
