% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{summary_dates}
\alias{summary_dates}
\title{covert all of the dates into one format for consistency
We can't use ":" to separate like in summary_numbers() because of %H:%M:%S in the column
Therefore we are using the regular expression to split using the first occurance of ":" and keeping the remaining values}
\usage{
summary_dates(data)
}
\description{
covert all of the dates into one format for consistency
We can't use ":" to separate like in summary_numbers() because of %H:%M:%S in the column
Therefore we are using the regular expression to split using the first occurance of ":" and keeping the remaining values
}
