% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{summary_logicals}
\alias{summary_logicals}
\title{count the number of TRUE and FALSE for logical type}
\usage{
summary_logicals(data)
}
\description{
count the number of TRUE and FALSE for logical type
}
