% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{tidy_summary}
\alias{tidy_summary}
\title{Function to print tidy version of "summary" function
This function converts input temperatures in Fahrenheit to Kelvin.}
\usage{
tidy_summary(data)
}
\arguments{
\item{summary_dcount}{gives distinct numbers}

\item{summary_numbers}{gives summary of numeric variables}

\item{summary_dates}{gives summary of date variables}

\item{join_summary}{uses summary_numbers and summary_dates to give a tidy summary}
}
\value{
distinct numbers, and basic statistics for each variable in the data
}
\description{
Function to print tidy version of "summary" function
This function converts input temperatures in Fahrenheit to Kelvin.
}
