% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{unique_keys_more}
\alias{unique_keys_more}
\title{Function that prints distinct values based on more than one group
 @param data name of the dataframe/datatable}
\usage{
unique_keys_more(data, ...)
}
\arguments{
\item{...}{a column name or a vector with multiple names}
}
\description{
Function that prints distinct values based on more than one group
 @param data name of the dataframe/datatable
}
