% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{unique_keys_one}
\alias{unique_keys_one}
\title{Function that prints distinct values based on one variable}
\usage{
unique_keys_one(data, ...)
}
\arguments{
\item{data}{name of the dataframe/datatable}

\item{...}{a column name or a vector with multiple names}
}
\description{
Function that prints distinct values based on one variable
}
