% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience.R
\name{z_score_transform}
\alias{z_score_transform}
\title{Standarize all the independent variables (convert the mean into 0 with standard deviation 1)
using `scale` function, which by default does z-score transformation}
\usage{
z_score_transform(data, ...)
}
\description{
Standarize all the independent variables (convert the mean into 0 with standard deviation 1)
using `scale` function, which by default does z-score transformation
}
